+++ demo/text.js
@@ -34,7 +34,7 @@ const input = add([
 		// The height of character
 		size: curSize,
 		lineSpacing: 8,
-		charSpacing: 4,
+		letterSpacing: 4,
 		// Transform each character for special effects
 		transform: (idx, ch) => ({
 			color: hsl2rgb((time() * 0.2 + idx * 0.1) % 1, 0.7, 0.8),
@@ -88,5 +88,23 @@ onKeyDown("down", () => {
 	curSize = Math.max(curSize - dt() * SIZE_SPEED, SIZE_MIN)
 	input.textSize = curSize
 })
+
+// Use this syntax and style option to style chunks of text, with CharTransformFunc.
+add([
+	text("[oh hi].green this is a [styled].wavy text", {
+		styles: {
+			"green": (idx, ch) => ({
+				color: GREEN,
+			}),
+			"wavy": (idx, ch) => ({
+				color: BLUE,
+				pos: vec2(0, wave(-4, 4, time() * 4 + idx * 0.5)),
+			}),
+		},
+	}),
+	pos(pad, height() - pad),
+	origin("botleft"),
+	scale(0.5),
+])
 
 // Check out https://kaboomjs.com#TextComp for everything text() offers